requirements.txt
fastapi
uvicorn
python-multipart
pydub
webrtcvad
faster-whisper
psycopg2-binary         # PostgreSQL driver
sqlalchemy              # ORM
python-dotenv           # load env vars
passlib[bcrypt]         # password hashing
bcrypt                  # low-level password hashing
python-jose[cryptography]  # JWT auth
pydantic                # request/response validation
